+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- PersistenceSettings.OnSaveError
+- PersistenceSettings.OnLoadError
 
 ## [1.0.0] - 2022-06-29
 ### Added